Terminado Crear Proyecto

+++ database/migrations/2019_12_04_223920_create_master_pages_table.php
@@ -28,7 +28,7 @@ class CreateMasterPagesTable extends Migration
             $table->string('seo_keywords_en')->nullable();
             $table->string('seo_image',100)->nullable();
 
-            $table->integer('index');   
+            $table->bigInteger('index');   
 
             $table->timestamps();
         });
