+++ database/migrations/2020_07_27_113124_create_members_table.php
@@ -17,7 +17,7 @@ class CreateMembersTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('image',100);
-            $table->integer('index');  
+            $table->bigInteger('index');  
             $table->timestamps();
         });
     }
