+++ database/migrations/2020_07_27_125600_create_type_department_blueprints_table.php
@@ -17,7 +17,7 @@ class CreateTypeDepartmentBlueprintsTable extends Migration
             $table->bigIncrements('id');
             $table->text('description_es');
             $table->text('description_en');
-            $table->integer('index'); 
+            $table->bigInteger('index'); 
             $table->string('image',100)->nullable(); 
 
             $table->bigInteger('type_department_id')->unsigned();
